requirements.txt
django==4.1.1
djangorestframework
psycopg2>=2.8
requests
django-cors-headers
Pillow